backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
httpx==0.25.2
sentence-transformers==2.7.0
numpy==1.24.3
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
retry==0.9.2

# Document Processing (Onyx-style)
nomic==0.0.1  # For nomic-ai embeddings
tiktoken==0.5.1  # For token counting
pypdf==3.17.4  # For PDF processing
python-docx==1.1.0  # For Word documents
openpyxl==3.1.2  # For Excel files
python-magic==0.4.27  # For file type detection
google-generativeai==0.3.2  # For Gemini LLM integration
qdrant-client==1.7.0  # For Qdrant vector databasem